feat: group join/creation view prototype
